Separate sync_test.go into util_test.go
